+++ src/main/java/com/telepathicgrunt/the_bumblezone/features/CaveSugarWaterfall.java
@@ -2,11 +2,11 @@ package com.telepathicgrunt.the_bumblezone.features;
 
 import com.mojang.serialization.Codec;
 import com.telepathicgrunt.the_bumblezone.fluids.BzFluids;
+import com.telepathicgrunt.the_bumblezone.tags.BZBlockTags;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
 import net.minecraft.util.Direction;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.registry.Registry;
 import net.minecraft.world.ISeedReader;
 import net.minecraft.world.gen.ChunkGenerator;
 import net.minecraft.world.gen.feature.Feature;
@@ -29,7 +29,7 @@ public class CaveSugarWaterfall extends Feature<NoFeatureConfig> {
         BlockPos.Mutable blockpos$Mutable = new BlockPos.Mutable().setPos(position);
         BlockState blockstate = world.getBlockState(blockpos$Mutable.up());
 
-        if (!blockstate.isSolid() || Registry.BLOCK.getKey(blockstate.getBlock()).getPath().contains("honey")) {
+        if (!blockstate.isSolid() || blockstate.isIn(BZBlockTags.HONEYCOMBS_THAT_FEATURES_CAN_CARVE)) {
             return false;
         } else {
             //checks if we are in the side of a wall with air exposed on one side
@@ -38,7 +38,7 @@ public class CaveSugarWaterfall extends Feature<NoFeatureConfig> {
             int neededNumberOfSides;
             blockstate = world.getBlockState(blockpos$Mutable.down());
 
-            if (blockstate.isSolid() && Registry.BLOCK.getKey(blockstate.getBlock()).getPath().contains("honey")) {
+            if (blockstate.isSolid() && blockstate.isIn(BZBlockTags.HONEYCOMBS_THAT_FEATURES_CAN_CARVE)) {
                 neededNumberOfSides = 3;
             } else if (blockstate.getBlock() == CAVE_AIR.getBlock()) {
                 neededNumberOfSides = 4;
@@ -49,7 +49,7 @@ public class CaveSugarWaterfall extends Feature<NoFeatureConfig> {
 
             for (Direction face : Direction.Plane.HORIZONTAL) {
                 blockstate = world.getBlockState(blockpos$Mutable.offset(face));
-                if (blockstate.isSolid() && Registry.BLOCK.getKey(blockstate.getBlock()).getPath().contains("honey")) {
+                if (blockstate.isSolid() && blockstate.isIn(BZBlockTags.HONEYCOMBS_THAT_FEATURES_CAN_CARVE)) {
                     ++numberOfSolidSides;
                 } else if (blockstate.getBlock() != CAVE_AIR.getBlock()) {
                     return false;